New ObjectGraph data from UC 11.3-11.8 on Android
This is the data from the remaining APKs listed in
https://github.com/mdittmer/web-apis/pull/47.

The same script was used, just by deleting the done APKs.

After removing the nginx proxy and using plain HTTP on port 8000,
debugging with wireshark showed that no packets were reaching the
server at all.

The problem appears to have been UC's data saver mode. It was disabled
in settings: ⚙ (更多设置) > 极速/省流 > 云端加速

Then everything worked fine.

The device was also re-flashed, so it's not a certain root cause.