New ObjectGraph data from Android 52-63 on Android
These tested versions are those listed on this blog:
https://chromereleases.googleblog.com/search/label/Chrome%20for%20Android+Stable%20updates

Posts that start with "Good news, everyone!" are the first release of
that version, and those were the versions tested.

Unfortuntely, the APKs for historical releases aren't available for
download from any public and offical location, and so aren't listed.

Script used:
```
adb uninstall com.android.chrome
for apk in *.apk; do
  read -p "Press enter to install and launch $apk"
  adb install -r "$apk"
  adb shell am start -a android.intent.action.VIEW -d https://wpt.foolip.org com.android.chrome
done
```

The same phone as for UC and QQ was used, a Nexus 5X.